Malware_Project_Final
The malware industry has developed into a well-resourced, high-value market in recent years. Multiplayer syndicates that are well-funded make significant investments in technologies and tools designed to circumvent conventional security, necessitating the development of countermeasures by anti-malware companies to identify and disable them. They cause actual financial and emotional harm to computer system users in the interim. The enormous amounts of data and files that need to be examined for any malicious intent provide one of the biggest issues anti-malware programs now face. We must be able to divide them into groups and recognize their various families in order to analyze and categorize such massive volumes of material effectively. Additionally, these grouping criteria may be used to identify new files that are discovered on computers as dangerous and belonging to a specific family. Microsoft is supporting the open-source development of efficient methods for classifying variant malware files into their respective families while also giving the data science community access to a previously unheard-of malware dataset and based on this data, I developed this project.